Add 64K timing monitoring, complexity control, and documentation
- Add TX_SLOW warnings when encode+TX cycle exceeds frame period
- Add TX end summary with enc_max, cycle_max, and overrun count
- Add UART commands (c/C) for runtime Opus complexity adjustment
- Default to complexity 2 for multi-packet modes (64K/96K)
  Testing shows: 0-1 same speed, 2 optimal, 3+ causes streaming defects
- Remove noisy OVERRUN@ debug logging (count still in TX summary)
- Update readme.md with 64K timing analysis section:
  - Actual timing breakdown from logic analyzer measurements
  - Opus complexity tuning table
  - Corrected margin values for 64K/96K modes (~5ms and ~0ms)
  - Serial command documentation for complexity control

+++ lora_transceiver/Inc/main.h
@@ -121,7 +121,7 @@ extern uint8_t frames_per_sec;
 #endif /* __MAIN_H */
 
 extern uint8_t lora_payload_length;
-extern uint8_t _bytes_per_frame;
+extern uint16_t _bytes_per_frame;
 extern uint8_t frame_length_bytes;
 extern uint8_t sf_at_500KHz;
